third_party/WebKit/LayoutTests/editing/deleting/delete-character-004.html

<div id="container">
<div id="sample" contenteditable="true"></div>
</div>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script>
function $(id) { return document.getElementById(id); }
var sample = $('sample');
sample.innerHTML = 'abc<br>';
sample.focus();
var range = document.createRange();
range.setStart(sample.firstChild, 3);
var selection = window.getSelection();
selection.removeAllRanges();
selection.addRange(range);
document.execCommand('Delete', false);

shouldBeEqualToString('sample.innerHTML', 'ab<br>');
shouldBeEqualToString('selection.type', 'Caret');
shouldBe('selection.focusNode', 'sample.firstChild');
shouldBe('selection.focusOffset', '2');

if (window.testRunner)
    $('container').outerHTML = '';
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
